app/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
instructor==0.3.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
reportlab
PyPDF2